STL_1_Dimensions/VERSION 2/requirements.txt
Flask
numpy
numpy-stl
